internal/llm: llm interfaces
Package llm defines the basic types involved with embedding using LLMs:
Embedder, EmbedDoc, and Vector.

It is possible this is too low-level an interface and that there should be
a DocDB interface as well.

It is also possible that storage.VectorDB should move into this package.

Copied from https://github.com/rsc/gaby/commit/0de1cd6.

Change-Id: Ia7a7c03b1cb332215c1adf3be98decaf9b133140